backend/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4 